Invoke api through method instead of endpoint
The current interface uses the endpoint to make the api request.
For example:

  api.trends.listThreads()

However that will not work if the endpoint contains a python
reserved word. A new interface was added:

  api.<method>(<endpoint>, <params>)

where method is the HTTP method of the request and endpoint is now
a string:

  api.get('trends.listThreads')

The legacy interface is still supported but I updated the docs since
the new method has better compatibility.